graphs/iso: Move Candidates to dedicated file.
